requirements.txt
cairosvg
pillow
# config is now loaded from config.json
